+++ adapters/tuya/__init__.py
@@ -4,6 +4,7 @@ from adapters.generic.temp_hum_sensor import TemperatureHumiditySensorAdapter
 from adapters.tuya.TS0002 import TS0002
 from adapters.tuya.TS0012 import TS0012
 from adapters.tuya.TS0041 import TS0041
+from adapters.tuya.TS0601 import TS0601
 
 
 tuya_adapters = {
@@ -14,4 +15,5 @@ tuya_adapters = {
     'TS0002': TS0002,                           # TuYa 2 gang switch
     'TS0012': TS0012,                           # Vrey VR-X711U
     'TS0041': TS0041,                           # TuYa Wireless switch with 1 button
+    'TS0601_thermostat': TS0601,                # TuYa Radiator valve with thermostat
 }